backend/requirements.txt
fastapi
uvicorn[standard]
fastmcp
langchain
langchain-community
langchain-google-genai
chromadb
pydantic
requests
python-dotenv
httpx